Use ChoicePointActivityListener in HeapOfActiveAtoms
