+++ src/main/java/at/ac/tuwien/kr/alpha/solver/heuristics/HeapOfActiveChoicePoints.java
@@ -26,7 +26,6 @@
 package at.ac.tuwien.kr.alpha.solver.heuristics;
 
 import at.ac.tuwien.kr.alpha.common.NoGood;
-import at.ac.tuwien.kr.alpha.solver.ChoiceInfluenceManager;
 import at.ac.tuwien.kr.alpha.solver.ChoiceManager;
 import org.apache.commons.collections4.MultiValuedMap;
 import org.apache.commons.collections4.multimap.HashSetValuedHashMap;
@@ -45,8 +44,6 @@ import static at.ac.tuwien.kr.alpha.common.Literals.atomOf;
  *
  */
 public class HeapOfActiveChoicePoints extends HeapOfActiveAtoms {
-	
-	private ChoiceManager choiceManager;
 
 	/**
 	 * Maps atoms to choice points representing bodies of rules in which the former atoms occur
@@ -55,9 +52,7 @@ public class HeapOfActiveChoicePoints extends HeapOfActiveAtoms {
 	protected final MultiValuedMap<Integer, Integer> atomsToChoicePoints = new HashSetValuedHashMap<>();
 	
 	public HeapOfActiveChoicePoints(int decayAge, double decayFactor, ChoiceManager choiceManager) {
-		super(decayAge, decayFactor);
-		this.choiceManager = choiceManager;
-		this.choiceManager.addChoicePointActivityListener(new HeuristicActivityListener());
+		super(decayAge, decayFactor, choiceManager);
 	}
 	
 	@Override
@@ -109,22 +104,6 @@ public class HeapOfActiveChoicePoints extends HeapOfActiveAtoms {
 	@Override
 	public String toString() {
 		return heap.toString();
-	}
-
-	private class HeuristicActivityListener implements ChoiceInfluenceManager.ActivityListener {
-
-		@Override
-		public void callbackOnChanged(int atom, boolean active) {
-			if (active && choiceManager.isActiveChoiceAtom(atom)) {
-				Double activity = activityScores.get(atom);
-				if (activity != null) {
-					/* if activity is null, probably the atom is still being buffered
-					   by DependencyDrivenVSIDSHeuristic and will get an initial activity
-					   when the buffer is ingested */
-					heap.add(atom);
-				}
-			}
-		}
 	}
 
 }
